pandaid is bigger than int32

+++ act/aCTDBPanda.py
@@ -47,7 +47,7 @@ class aCTDBPanda(aCTDB):
         id INTEGER PRIMARY KEY AUTO_INCREMENT,
         modified TIMESTAMP,
         created TIMESTAMP,
-        pandaid integer, 
+        pandaid bigint, 
         pandajob text,
         siteName VARCHAR(255),
         prodSourceLabel VARCHAR(255),
@@ -77,7 +77,7 @@ class aCTDBPanda(aCTDB):
         
         str="""
         create table pandaarchive (
-        pandaid integer, 
+        pandaid bigint, 
         siteName VARCHAR(255),
         actpandastatus VARCHAR(255),
         startTime TIMESTAMP DEFAULT 0,
